requirements.txt
python-dotenv==1.1.0
Requests==2.32.4